add exception if user not found for deleting

+++ src/main/java/com/example/servicesprovider/base/service/impl/BaseServiceImpl.java
@@ -55,6 +55,10 @@ public abstract class BaseServiceImpl<E extends BaseModel<ID>, ID extends Serial
 
     @Override
     public void deleteById(ID id) {
-        repository.deleteById(id);
+        Optional<E> e = repository.findById(id);
+        if (e.isPresent()) {
+            repository.deleteById(id);
+        }
+        throw new EntityNotFoundException("Model does not exist");
     }
 }
